// src/interface/user.interface.ts
export interface userData {
    user: {
        first_name: string;
        second_name: string;
        last_name: string;
        second_surname: string;
        CI: string;
        gender: string;
        career: string;
        cohort: string;
        direction: string;
        phone: string;
        status: boolean;
        study_location: string;
        _id: string;
    }

}

export interface user {
    first_name: string;
    second_name: string;
    last_name: string;
    second_surname: string;
    CI: string;
    gender: string;
    career: string;
    cohort: string;
    direction: string;
    phone: string;
    status: boolean;
    study_location: string;
    _id: string;
}